splitted AmountCalculator in seperated classes
