+++ src/main/java/io/konik/calculation/InvoiceMonetarySummationAndTradeTaxCompleter.java
@@ -1,46 +1,47 @@
 package io.konik.calculation;
 
+import static io.konik.validation.AmountCalculator.recalculate;
+import java.util.List;
+import javax.validation.constraints.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import io.konik.validation.RecalculationResult;
+import io.konik.validation.TaxAggregator;
 import io.konik.zugferd.Invoice;
 import io.konik.zugferd.entity.trade.MonetarySummation;
 import io.konik.zugferd.entity.trade.TradeTax;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.validation.constraints.NotNull;
-import java.util.List;
-
-import static io.konik.validation.AmountCalculator.*;
 
 /**
- * Recalculates invoice settlement {@link MonetarySummation} and replaces it in the {@link Invoice} object.
+ * Recalculates invoice settlement {@link MonetarySummation} and replaces it in the {@link Invoice}
+ * object.
  */
 public final class InvoiceMonetarySummationAndTradeTaxCompleter implements Correction<Invoice> {
 
-	protected static Logger log = LoggerFactory.getLogger(InvoiceMonetarySummationAndTradeTaxCompleter.class);
+  protected static Logger log =
+      LoggerFactory.getLogger(InvoiceMonetarySummationAndTradeTaxCompleter.class);
 
-	@Override
-	public Invoice correct(@NotNull final Invoice invoice) {
+  @Override
+  public Invoice correct(@NotNull final Invoice invoice) {
 
-		log.debug("Running InvoiceMonetarySummationCompleter...");
+    log.debug("Running InvoiceMonetarySummationCompleter...");
 
-		if (invoice.getTrade() != null && invoice.getTrade().getSettlement() != null) {
-			RecalculationResult result = recalculate(invoice);
-			MonetarySummation monetarySummation = result.getMonetarySummation();
-			invoice.getTrade().getSettlement().setMonetarySummation(monetarySummation);
+    if (invoice.getTrade() != null && invoice.getTrade().getSettlement() != null) {
+      RecalculationResult result = recalculate(invoice);
+      MonetarySummation monetarySummation = result.getMonetarySummation();
+      invoice.getTrade().getSettlement().setMonetarySummation(monetarySummation);
 
-			TaxAggregator taxAggregator = result.getTaxAggregator();
-			List<TradeTax> taxes = taxAggregator.generateTradeTaxList(
-					invoice.getTrade().getSettlement().getCurrency(),
-					invoice.getTrade().getSettlement().getTradeTax()
-			);
+      TaxAggregator taxAggregator = result.getTaxAggregator();
+      List<TradeTax> taxes =
+          taxAggregator.generateTradeTaxList(invoice.getTrade().getSettlement().getCurrency(),
+              invoice.getTrade().getSettlement().getTradeTax());
 
-			invoice.getTrade().getSettlement().getTradeTax().clear();
+      invoice.getTrade().getSettlement().getTradeTax().clear();
 
-			for (TradeTax tradeTax : taxes) {
-				invoice.getTrade().getSettlement().addTradeTax(tradeTax);
-			}
+      for (TradeTax tradeTax : taxes) {
+        invoice.getTrade().getSettlement().addTradeTax(tradeTax);
+      }
 
-		}
-		return invoice;
-	}
+    }
+    return invoice;
+  }
 }
