+++ src/main/java/io/konik/validation/AmountCalculator.java
@@ -19,30 +19,19 @@ package io.konik.validation;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.google.common.base.Function;
 import com.neovisionaries.i18n.CurrencyCode;
 import io.konik.util.Amounts;
 import io.konik.util.Items;
 import io.konik.util.MonetarySummations;
 import io.konik.zugferd.Invoice;
-import io.konik.zugferd.entity.AllowanceCharge;
 import io.konik.zugferd.entity.AppliedTax;
-import io.konik.zugferd.entity.GrossPrice;
 import io.konik.zugferd.entity.LogisticsServiceCharge;
 import io.konik.zugferd.entity.SpecifiedAllowanceCharge;
-import io.konik.zugferd.entity.Tax;
 import io.konik.zugferd.entity.trade.MonetarySummation;
 import io.konik.zugferd.entity.trade.Settlement;
-import io.konik.zugferd.entity.trade.TradeTax;
 import io.konik.zugferd.entity.trade.item.Item;
 import io.konik.zugferd.entity.trade.item.ItemTax;
 import io.konik.zugferd.entity.trade.item.SpecifiedAgreement;
@@ -118,7 +107,8 @@ public final class AmountCalculator {
 
     appendTaxFromInvoiceServiceCharge(settlement, taxAggregator);
 
-    monetarySummation.setTaxBasisTotal(new Amount(taxAggregator.calculateTaxBasis(), currency));
+    monetarySummation
+        .setTaxBasisTotal(new Amount(taxAggregator.calculateTaxBasis(items), currency));
     monetarySummation.setTaxTotal(new Amount(taxAggregator.calculateTaxTotal(), currency));
 
     monetarySummation.setGrandTotal(
@@ -249,326 +239,6 @@ public final class AmountCalculator {
   private static void assertNotNull(final Item item) {
     if (item == null) {
       throw new IllegalArgumentException("Item cannot be null");
-    }
-  }
-
-
-  /**
-   * Helper class for calculating {@link Item}'s line total.
-   */
-  static final class ItemLineTotalCalculator implements Function<Item, Amount> {
-
-    @Nullable
-    @Override
-    public Amount apply(@Nullable Item item) {
-      Amount originLineTotal = null;
-
-      if (item != null && item.getSettlement() != null
-          && item.getSettlement().getMonetarySummation() != null) {
-        originLineTotal = Amounts.copy(item.getSettlement().getMonetarySummation().getLineTotal());
-      }
-
-      if (item == null || item.getDelivery() == null || item.getAgreement() == null) {
-        return originLineTotal;
-      }
-
-      if (item.getAgreement().getNetPrice() == null) {
-        return originLineTotal;
-      }
-
-      BigDecimal quantity =
-          item.getDelivery().getBilled() != null ? item.getDelivery().getBilled().getValue()
-              : BigDecimal.ZERO;
-      Amount amount = item.getAgreement().getNetPrice().getChargeAmount();
-
-      log.debug("Line total formula: {} (net price) x {} (quantity)", amount, quantity);
-
-      return Amounts.multiply(amount, quantity);
-    }
-  }
-
-  /**
-   * Helper class for calculating {@link Item}'s tax total.
-   */
-  static final class ItemTaxTotalCalculator implements Function<Item, Amount> {
-
-    private final Amount lineTotal;
-
-    public ItemTaxTotalCalculator(final Amount lineTotal) {
-      this.lineTotal = lineTotal;
-    }
-
-    @Override
-    public Amount apply(@Nullable Item item) {
-      CurrencyCode currency = lineTotal.getCurrency();
-      Amount taxTotal = Amounts.zero(currency);
-
-      if (item != null && item.getSettlement() != null
-          && item.getSettlement().getTradeTax() != null) {
-        for (ItemTax tax : item.getSettlement().getTradeTax()) {
-          BigDecimal taxRate =
-              tax.getPercentage().divide(BigDecimal.valueOf(100), 2, BigDecimal.ROUND_HALF_UP)
-                  .setScale(2, RoundingMode.HALF_UP);
-          BigDecimal taxValue =
-              lineTotal.getValue().multiply(taxRate).setScale(2, RoundingMode.HALF_UP);
-
-          taxTotal = Amounts.add(taxTotal, new Amount(taxValue, currency));
-        }
-      }
-      return taxTotal;
-    }
-  }
-
-  /**
-   * Calculates total {@link io.konik.zugferd.entity.AllowanceCharge} for given {@link Item}.
-   */
-  public static final class ItemTotalAllowanceChargeCalculator implements Function<Item, Amount> {
-
-    private final CurrencyCode currencyCode;
-
-    public ItemTotalAllowanceChargeCalculator(CurrencyCode currencyCode) {
-      this.currencyCode = currencyCode;
-    }
-
-    @Override
-    public Amount apply(@Nullable Item item) {
-      Amount totalAllowanceCharge = Amounts.zero(currencyCode);
-      BigDecimal quantity = BigDecimal.ONE;
-
-      if (item != null && item.getDelivery() != null && item.getDelivery().getBilled() != null) {
-        quantity = item.getDelivery().getBilled().getValue();
-      }
-
-      if (item != null && item.getAgreement() != null
-          && item.getAgreement().getGrossPrice() != null) {
-        GrossPrice grossPrice = item.getAgreement().getGrossPrice();
-
-        if (grossPrice.getAllowanceCharges() != null
-            && !grossPrice.getAllowanceCharges().isEmpty()) {
-          for (AllowanceCharge charge : grossPrice.getAllowanceCharges()) {
-            BigDecimal chargeValue = charge.getActual().getValue();
-            if (charge.isDiscount()) {
-              chargeValue = chargeValue.negate();
-            }
-            Amount amount = new Amount(chargeValue.multiply(quantity), currencyCode);
-            totalAllowanceCharge = Amounts.add(totalAllowanceCharge, amount);
-          }
-
-          totalAllowanceCharge =
-              Amounts.setPrecision(totalAllowanceCharge, 2, RoundingMode.HALF_UP);
-        }
-      }
-
-      return totalAllowanceCharge;
-    }
-  }
-
-  public static final class InvoiceAllowanceTotalCalculator
-      implements Function<Settlement, Amount> {
-    @Nullable
-    @Override
-    public Amount apply(@Nullable Settlement settlement) {
-      if (settlement == null || settlement.getAllowanceCharge() == null) {
-        return null;
-      }
-
-      BigDecimal chargeValue = BigDecimal.ZERO;
-
-      for (SpecifiedAllowanceCharge charge : settlement.getAllowanceCharge()) {
-        if (charge.isDiscount()) {
-          chargeValue = chargeValue.add(charge.getActual().getValue());
-        }
-      }
-      Amount amount = new Amount(chargeValue, settlement.getCurrency());
-
-      log.debug("Invoice allowance total = {}", amount);
-
-      return amount;
-    }
-  }
-
-  public static final class InvoiceChargeTotalCalculator implements Function<Settlement, Amount> {
-    @Nullable
-    @Override
-    public Amount apply(@Nullable Settlement settlement) {
-      if (settlement == null || settlement.getAllowanceCharge() == null) {
-        return null;
-      }
-
-      BigDecimal chargeValue = BigDecimal.ZERO;
-
-      for (SpecifiedAllowanceCharge charge : settlement.getAllowanceCharge()) {
-        if (charge.isSurcharge()) {
-          chargeValue = chargeValue.add(charge.getActual().getValue());
-        }
-      }
-
-      for (LogisticsServiceCharge charge : settlement.getServiceCharge()) {
-        chargeValue = chargeValue.add(charge.getAmount().getValue());
-      }
-
-      Amount amount = new Amount(chargeValue, settlement.getCurrency());
-
-      log.debug("Invoice charge total = {}", amount);
-
-      return amount;
-    }
-  }
-
-  public static final class ItemTaxExtractor implements Function<Item, ItemTax> {
-    @Nullable
-    @Override
-    public ItemTax apply(@Nullable Item item) {
-      if (item == null || item.getSettlement() == null) {
-        return null;
-      }
-
-      if (item.getSettlement().getTradeTax().isEmpty()) {
-        return null;
-      }
-
-      return item.getSettlement().getTradeTax().size() > 1
-          ? item.getSettlement().getTradeTax().get(1)
-          : item.getSettlement().getTradeTax().get(0);
-    }
-  }
-
-  /**
-   * Helper class for aggregating tax information and calculating tax basis and tax total values.
-   */
-  public static final class TaxAggregator {
-
-    private static final int PRECISION = 2;
-    private RoundingMode roundingMode;
-
-    private final ConcurrentMap<Key, BigDecimal> map = new ConcurrentHashMap<Key, BigDecimal>();
-
-    public TaxAggregator() {
-      roundingMode = RoundingMode
-          .valueOf(System.getProperty(this.getClass().getName() + ".RoundingMode", "HALF_UP"));
-    }
-
-    public void add(Tax tax, BigDecimal amount) {
-      Key key = Key.create(tax);
-      map.putIfAbsent(key, BigDecimal.ZERO);
-      map.put(key, map.get(key).add(amount));
-    }
-
-    public BigDecimal getTaxBasisForTaxPercentage(final BigDecimal percentage) {
-      BigDecimal value = BigDecimal.ZERO;
-      for (Key key : map.keySet()) {
-        if (percentage.equals(key.getPercentage())) {
-          value = value.add(map.get(key));
-        }
-      }
-      return value;
-    }
-
-    public BigDecimal calculateTaxBasis() {
-      log.debug("Recalculating tax basis for tax percentages: {}",
-          Arrays.toString(map.keySet().toArray()));
-      BigDecimal taxBasis = BigDecimal.ZERO;
-      for (BigDecimal amount : map.values()) {
-        taxBasis = taxBasis.add(amount);
-      }
-
-      log.debug("Recalculated tax basis = {}", taxBasis);
-
-      return taxBasis;
-    }
-
-    public BigDecimal calculateTaxTotal() {
-      log.debug("Calculating tax total...");
-      BigDecimal taxTotal = BigDecimal.ZERO;
-      for (Map.Entry<Key, BigDecimal> entry : map.entrySet()) {
-        BigDecimal percentage = entry.getKey().getPercentage();
-        BigDecimal value = entry.getValue();
-        BigDecimal taxAmount = calculateTaxAmount(percentage, value);
-
-        log.debug("===> {} x {}% = {}", value, percentage, taxAmount);
-
-        taxTotal = taxTotal.add(taxAmount);
-      }
-
-      log.debug("Recalculated tax total = {}", taxTotal);
-
-      return taxTotal;
-    }
-
-    public List<TradeTax> generateTradeTaxList(final CurrencyCode currencyCode,
-        final List<TradeTax> previousList) {
-      List<TradeTax> taxes = new LinkedList<TradeTax>();
-
-      for (Key key : map.keySet()) {
-        TradeTax tradeTax = new TradeTax();
-        tradeTax.setType(key.getCode());
-        tradeTax.setCategory(key.getCategory());
-        tradeTax.setPercentage(key.getPercentage());
-
-        BigDecimal basis = map.get(key);
-        BigDecimal calculated = calculateTaxAmount(key.getPercentage(), basis);
-
-        tradeTax.setBasis(new Amount(basis, currencyCode));
-        tradeTax.setCalculated(new Amount(calculated, currencyCode));
-
-        TradeTax existing = null;
-        if (previousList != null) {
-          for (TradeTax current : previousList) {
-            if (tradeTax.getType().equals(current.getType())
-                && tradeTax.getCategory().equals(current.getCategory())
-                && tradeTax.getPercentage().equals(current.getPercentage())) {
-              existing = current;
-              break;
-            }
-          }
-        }
-
-        if (existing != null) {
-          tradeTax.setExemptionReason(existing.getExemptionReason());
-
-          if (existing.getAllowanceCharge() != null) {
-            tradeTax.setAllowanceCharge(new Amount(existing.getAllowanceCharge().getValue(),
-                existing.getAllowanceCharge().getCurrency()));
-          }
-
-          if (existing.getLineTotal() != null) {
-            tradeTax.setLineTotal(new Amount(existing.getLineTotal().getValue(),
-                existing.getLineTotal().getCurrency()));
-          }
-        }
-
-        taxes.add(tradeTax);
-      }
-
-      return taxes;
-    }
-
-    public BigDecimal calculateTaxAmount(final BigDecimal percentage, final BigDecimal value) {
-      return value.multiply(percentage.divide(BigDecimal.valueOf(100))).setScale(PRECISION,
-          roundingMode);
-    }
-
-    @Override
-    public String toString() {
-      return "TaxAggregator{" + "map=" + map + '}';
-    }
-  }
-
-  public static final class RecalculationResult {
-    private final MonetarySummation monetarySummation;
-    private final TaxAggregator taxAggregator;
-
-    public RecalculationResult(MonetarySummation monetarySummation, TaxAggregator taxAggregator) {
-      this.monetarySummation = monetarySummation;
-      this.taxAggregator = taxAggregator;
-    }
-
-    public MonetarySummation getMonetarySummation() {
-      return monetarySummation;
-    }
-
-    public TaxAggregator getTaxAggregator() {
-      return taxAggregator;
     }
   }
 }
